Fix crash in 32 bit arch (#22)
* Fix crash in 32 bit arch

* Fix TravisCI script

+++ Sources/ScryptParams.swift
@@ -72,7 +72,7 @@ public struct ScryptParams {
     ///
     /// - Returns: a `ValidationError` or `nil` if the parameters are valid.
     public func validate() -> ValidationError? {
-        if desiredKeyLength > ((1 << 32) - 1) * 32 {
+        if desiredKeyLength > ((1 << 32 as Int64) - 1 as Int64) * 32 {
             return ValidationError.desiredKeyLengthTooLarge
         }
         if UInt64(r) * UInt64(p) >= (1 << 30) {
